fix(backend): 优化删除题目题库关联逻辑
- 在删除题目题库关联时，先判断关联是否存在，避免不必要的删除操作
- 添加空格优化代码格式

+++ mianshiya-backend/src/main/java/com/allen/mianshiya/service/impl/QuestionServiceImpl.java
@@ -169,7 +169,7 @@ public class QuestionServiceImpl extends ServiceImpl<QuestionMapper, Question> i
                 Set<Long> questionIdSet = questionList.stream().map(Question::getId).collect(Collectors.toSet());
                 // 复用原有题目表的查询条件
                 queryWrapper.in("id", questionIdSet);
-            }else {
+            } else {
                 return new Page<>(current, size);
             }
         }
@@ -381,8 +381,11 @@ public class QuestionServiceImpl extends ServiceImpl<QuestionMapper, Question> i
             // 移除题目题库关系
             LambdaQueryWrapper<QuestionBankQuestion> lambdaQueryWrapper = Wrappers.lambdaQuery(QuestionBankQuestion.class)
                     .eq(QuestionBankQuestion::getQuestionId, questionId);
-            result = questionBankQuestionService.remove(lambdaQueryWrapper);
-            ThrowUtils.throwIf(!result, ErrorCode.OPERATION_ERROR, "删除题目题库关联失败");
+            List<QuestionBankQuestion> list = questionBankQuestionService.list(lambdaQueryWrapper);
+            if (CollUtil.isNotEmpty(list)) {
+                result = questionBankQuestionService.remove(lambdaQueryWrapper);
+                ThrowUtils.throwIf(!result, ErrorCode.OPERATION_ERROR, "删除题目题库关联失败");
+            }
         }
         return true;
     }
